server/server/requirements.txt
Django==3.0
gunicorn==19.10.0
django-cors-headers==3.2.1
django_otp==0.8.1
PyJWT==1.7.1
djangorestframework-simplejwt
django-extensions==2.2.8
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
filetype==1.0.7
psycopg2==2.8.4
twilio==6.43.0
djangorestframework==3.11.0
django-filter==2.4.0
boto3==1.13.16
django-storages==1.9.1
Pillow==7.1.2
PyPDF2==1.26.0
camelot-py==0.9.0
opencv-python==4.5.2.54
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22.1